+++ docs/CHANGELOG.md
@@ -15,6 +15,14 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.25.6] - 12-09-23
+### Fixed
+- Fixed Nodes and Edges step
+- Fixed issues with regex
+### Added
+- Mock Graph Generation Step
+- Regex paterns from CDF API documentation
+
 ## [0.25.5] - 5-09-23
 ### Added
 - Support for upload of various RDF formats to `NeatGraph` store